[0.7.0] refactor(borsh-rs): Avoid silent integer casts since they can lead to hidden security issues (#75)
Use `TryFrom` instead of `as u8` to make use of Rust strong type system. Casting integers is unsafe, we should not silently truncate the numbers. Also, in many of these cases, Rust will see that on the target platform `usize` is >= `u32` and should drop away those panic branches completely.

# Test plan

* `cargo test` passes without changes to the tests suite