fix CI
CI no longer works for 1.40, due to some dependency no longer being
compatible with it. Let's scale down our ambitious with respect to the
range of Rust version we intend to support to just the current stable.
Of course it would be *nice* to support more things, but we clearly
failing behind on that, with the current CI being broken on master.