de: add EnumExt trait with deserialize_variant method
Introduce de::EnumExt trait with deserialize_variant method which
takes variant number as an argument rather than reading it from the
input.

This trait is derived for all enums which use BorshDeserialize derive.

It is useful when customer wants to perform some validation after the
variant is known but before the variant is fully deserialised.