deserialize: prevent unbound allocation
With a crafted input, a malformed data of just four bytes may lead to
allocation of 4 GiB.  This may lead to OOM on constrained environments
(e.g. inside of a container with memory limits set or on 32-bit
machines).

(This once again shows Rust desperately needs full specialisation.  We
may dream of time that it ever comes…)