Enable `derive` when `schema` feature is enabled
Currently, enabling `schema` without enabling `derive` feature causes
a compilation error.  The error describes the solution which is great but
it raises the question that if the code knows how to solve the problem
why doesn’t it do it automatically? This feels user-hostile.

Change the features such that `schema` enables `derive` eliminating that
error condition.